Coordinate many endpoints with a single Tor
Here we create a DeferredDispatcher class to help coordinate
using a single instance of Tor. Each endpoint listen method
returns a deferred that it also gives to torDeferredDispatcher,
the txtorcon module scoped instance of DeferredDispatcher.

When Tor is finished booting then it fires the torDeferredDispatcher
with a (control_port, socks_port) 2-tuple. torDeferredDispatcher fires
all accumulated deferreds with the 2-tuple value. When in the "fired" state,
all deferreds passed to the DeferredDispatcher will be fired immediately.

In this way the code calling a listen on it's Tor Hidden Service endpoint
doesn't need to know if the Tor process is booted up already or not.
It simply receives a deferred which will eventually fire with a TorOnionListeningPort
which implements IListeningPort.

This commit also fixes the TCP port selection logic. We now use the kernel's
selection of TCP ports by specifying a TCP port of zero.